README
In which we create a simple game where you take a generic fantasy hero and make them fight a generic fantasy monster. Originally this was going to have a map, and I actually did get the map to render. But I couldn't figure out a way to make that interesting beyond just randomly wandering around the map until you bumped into the critter and getting the monster to a random spot was boring and fiddly so I decided to focus on the combat instead. Warning: I was tired and hungry when I wrote the comments, so expect some sass.

Okay, a lot of sass.

Combat here is simple: You have four stats: Health, Attack, Healing, and Defense
Health is your total HP. If it reaches 0 you die
Attack is your damage output.
Healing is your ability to heal yourself after taking damage.
Defense is your resistance to damage.
Damage is calculated by rolling your Attack score and then subtracting the result of rolling the target's Defense score
Everything is determined by rolls, with the score being the maximum result. Think of it as being a dWhatever dice
After the player attacks the monster attacks back. Rinse, repeat, until somebody dies.

There is a secret easter egg class and monster, Jedi and Moose. They're both ridiculously overpowered.